# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Grinch API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionAuth: []
paths:
  /machines:
    get:
      operationId: listMachines
      tags:
        - machines
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/UserIdFilter'
      responses:
        '200':
          description: Machine list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /machines/{id}:
    get:
      operationId: getMachine
      tags:
        - machines
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Machine detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteMachine
      tags:
        - machines
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Machine deleted.
        default:
          $ref: '#/components/responses/Problem'
...
  /groups:
    get:
      operationId: listGroups
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Group list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createGroup
      tags:
        - groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
      responses:
        '201':
          description: Group created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        default:
          $ref: '#/components/responses/Problem'
  /groups/{id}:
    get:
      operationId: getGroup
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Group detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchGroup
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupPatchRequest'
      responses:
        '200':
          description: Group updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteGroup
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Group deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /group-memberships:
    get:
      operationId: listGroupMemberships
      tags:
        - group-memberships
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/GroupIdFilter'
        - $ref: '#/components/parameters/UserIdFilter'
        - $ref: '#/components/parameters/MachineIdFilter'
      responses:
        '200':
          description: Group membership list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMembershipListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createGroupMembership
      tags:
        - group-memberships
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMembershipCreateRequest'
      responses:
        '201':
          description: Group membership created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMembership'
        default:
          $ref: '#/components/responses/Problem'
  /group-memberships/{id}:
    get:
      operationId: getGroupMembership
      tags:
        - group-memberships
      parameters:
        - $ref: '#/components/parameters/GroupMembershipId'
      responses:
        '200':
          description: Group membership detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMembership'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteGroupMembership
      tags:
        - group-memberships
      parameters:
        - $ref: '#/components/parameters/GroupMembershipId'
      responses:
        '204':
          description: Group membership deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
...
  /execution-events:
    get:
      operationId: listExecutionEvents
      tags:
        - execution-events
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/MachineIdFilter'
        - $ref: '#/components/parameters/UserIdFilter'
        - $ref: '#/components/parameters/ExecutableIdFilter'
      responses:
        '200':
          description: Execution event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionEventListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /execution-events/{id}:
    get:
      operationId: getExecutionEvent
      tags:
        - execution-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Execution event detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteExecutionEvent
      tags:
        - execution-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Execution event deleted.
        default:
          $ref: '#/components/responses/Problem'
  /file-access-events:
    get:
      operationId: listFileAccessEvents
      tags:
        - file-access-events
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/MachineIdFilter'
        - $ref: '#/components/parameters/ExecutableIdFilter'
      responses:
        '200':
          description: File access event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileAccessEventListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /file-access-events/{id}:
    get:
      operationId: getFileAccessEvent
      tags:
        - file-access-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: File access event detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileAccessEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteFileAccessEvent
      tags:
        - file-access-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: File access event deleted.
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: grinch_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        format: int32
        minimum: 1
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        format: int32
        minimum: 0
    Search:
      name: search
      in: query
      schema:
        type: string
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Sort field name.
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
    UserIdFilter:
      name: user_id
      in: query
      schema:
        type: string
        format: uuid
    MachineIdFilter:
      name: machine_id
      in: query
      schema:
        type: string
        format: uuid
    ExecutableIdFilter:
      name: executable_id
      in: query
      schema:
        type: string
        format: uuid
    GroupIdFilter:
      name: group_id
      in: query
      schema:
        type: string
        format: uuid
...
    GroupMembershipId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Problem response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - code
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        code:
          type: string
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...
    Source:
      type: string
      enum:
        - local
        - entra
    MemberKind:
      type: string
      enum:
        - user
        - machine
    GroupMembershipKind:
      type: string
      enum:
        - actual
        - effective
...
    EventDecision:
      type: string
      enum:
        - unknown
        - allow_unknown
        - allow_binary
        - allow_certificate
        - allow_scope
        - allow_team_id
        - allow_signing_id
        - allow_cd_hash
        - block_unknown
        - block_binary
        - block_certificate
        - block_scope
        - block_team_id
        - block_signing_id
        - block_cd_hash
        - bundle_binary
...
    MachineSummary:
      type: object
      required:
        - id
        - serial_number
        - hostname
        - model_identifier
        - os_version
        - santa_version
        - primary_user
        - last_seen_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        serial_number:
          type: string
        hostname:
          type: string
        model_identifier:
          type: string
        os_version:
          type: string
        santa_version:
          type: string
        primary_user:
          type: string
        primary_user_id:
          type: string
          format: uuid
          nullable: true
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Machine:
      type: object
      required:
        - id
        - serial_number
        - hostname
        - model_identifier
        - os_version
        - os_build
        - santa_version
        - primary_user
        - request_clean_sync
        - last_seen_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        serial_number:
          type: string
        hostname:
          type: string
        model_identifier:
          type: string
        os_version:
          type: string
        os_build:
          type: string
        santa_version:
          type: string
        primary_user:
          type: string
        primary_user_id:
          type: string
          format: uuid
          nullable: true
        request_clean_sync:
          type: boolean
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Group:
      type: object
      required:
        - id
        - name
        - description
        - source
        - member_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        source:
          $ref: '#/components/schemas/Source'
        member_count:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GroupCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    GroupPatchRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        description:
          type: string
    GroupMembershipGroup:
      type: object
      required:
        - id
        - name
        - source
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        source:
          $ref: '#/components/schemas/Source'
    GroupMembershipMember:
      type: object
      required:
        - kind
        - id
      properties:
        kind:
          $ref: '#/components/schemas/MemberKind'
        id:
          type: string
          format: uuid
        name:
          type: string
    GroupMembership:
      type: object
      required:
        - id
        - kind
        - group
        - member
        - created_at
        - updated_at
      properties:
        id:
          type: string
        kind:
          $ref: '#/components/schemas/GroupMembershipKind'
        group:
          $ref: '#/components/schemas/GroupMembershipGroup'
        member:
          $ref: '#/components/schemas/GroupMembershipMember'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GroupMembershipCreateRequest:
      type: object
      required:
        - group_id
        - member_kind
        - member_id
      properties:
        group_id:
          type: string
          format: uuid
        member_kind:
          $ref: '#/components/schemas/MemberKind'
        member_id:
          type: string
          format: uuid
...
    SigningChainEntry:
      type: object
      required:
        - common_name
        - organization
        - organizational_unit
        - sha256
        - valid_from
        - valid_until
      properties:
        common_name:
          type: string
        organization:
          type: string
        organizational_unit:
          type: string
        sha256:
          type: string
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
...
    ExecutionEventSummary:
      type: object
      required:
        - id
        - machine_id
        - executable_id
        - decision
        - file_path
        - file_name
        - signing_id
        - created_at
      properties:
        id:
          type: string
          format: uuid
        machine_id:
          type: string
          format: uuid
        executable_id:
          type: string
          format: uuid
        decision:
          $ref: '#/components/schemas/EventDecision'
        file_path:
          type: string
        file_name:
          type: string
        signing_id:
          type: string
        occurred_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ExecutionEvent:
      type: object
      required:
        - id
        - machine_id
        - executable_id
        - decision
        - file_path
        - file_name
        - file_sha256
        - file_bundle_id
        - file_bundle_path
        - signing_id
        - team_id
        - cdhash
        - executing_user
        - logged_in_users
        - current_sessions
        - signing_chain
        - entitlements
        - created_at
      properties:
        id:
          type: string
          format: uuid
        machine_id:
          type: string
          format: uuid
        executable_id:
          type: string
          format: uuid
        decision:
          $ref: '#/components/schemas/EventDecision'
        file_path:
          type: string
        file_name:
          type: string
        file_sha256:
          type: string
        file_bundle_id:
          type: string
        file_bundle_path:
          type: string
        signing_id:
          type: string
        team_id:
          type: string
        cdhash:
          type: string
        executing_user:
          type: string
        logged_in_users:
          type: array
          items:
            type: string
        current_sessions:
          type: array
          items:
            type: string
        signing_chain:
          type: array
          items:
            $ref: '#/components/schemas/SigningChainEntry'
        entitlements:
          type: object
          additionalProperties: true
        occurred_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    FileAccessEvent:
      type: object
      required:
        - id
        - machine_id
        - rule_version
        - rule_name
        - target
        - decision
        - file_name
        - file_sha256
        - signing_id
        - team_id
        - cdhash
        - process_chain
        - created_at
      properties:
        id:
          type: string
          format: uuid
        machine_id:
          type: string
          format: uuid
        executable_id:
          type: string
          format: uuid
          nullable: true
        rule_version:
          type: string
        rule_name:
          type: string
        target:
          type: string
        decision:
          $ref: '#/components/schemas/FileAccessDecision'
        file_name:
          type: string
        file_sha256:
          type: string
        signing_id:
          type: string
        team_id:
          type: string
        cdhash:
          type: string
        process_chain:
          type: array
          items:
            $ref: '#/components/schemas/FileAccessEventProcess'
        occurred_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    MachineListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/MachineSummary'
...
    GroupListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    GroupMembershipListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/GroupMembership'
...
    ExecutionEventListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionEventSummary'
    FileAccessEventListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/FileAccessEventSummary'